Tracer: Make locals width and height size_t to avoid implicit conversions from int

+++ src/RayTracer.cpp
@@ -19,8 +19,8 @@ RayTracer::RayTracer()
 // traceScene it through the scene and write computed color to buffer (dynamically scheduled in parallel using openMp)
 void RayTracer::traceScene(const Camera& camera, const Scene& scene, FrameBuffer& frameBuffer) const {
     // use ints for indexing since size_t is not supported by openMp loop parallelization macros
-    const int width        = static_cast<int>(frameBuffer.width());
-    const int height       = static_cast<int>(frameBuffer.height());
+    const size_t width     = static_cast<int>(frameBuffer.width());
+    const size_t height    = static_cast<int>(frameBuffer.height());
     const float invWidth   = 1.00f / width;
     const float invHeight  = 1.00f / height;
     const Vec3 eyePosition = camera.position();
@@ -31,7 +31,7 @@ void RayTracer::traceScene(const Camera& camera, const Scene& scene, FrameBuffer
         const Vec3 viewportPosition{ (col + 0.50f) * invWidth, (row + 0.50f) * invHeight, 0.00f };
         const Ray primaryRay = camera.viewportPointToRay(viewportPosition);
         const Color pixelColor = traceRay(camera, scene, primaryRay, 0);
-        frameBuffer.setPixel(height - 1 - row, col, pixelColor);  // invert y since viewport and row start opposite
+        frameBuffer.setPixel(height - 1 - row, col, pixelColor);  // invert y (since viewport and row start opposite)
     }
 }
 
@@ -145,6 +145,7 @@ Color RayTracer::computeSpecularColor(const Intersection& intersection, const IL
     const float strengthAtCamAngle = Math::max(0.00f, Math::dot(intersection.normal, halfwayVec));
     return Math::pow(strengthAtCamAngle, surfaceMaterial.shininess()) * surfaceMaterial.specularColor();
 }
+
 
 std::ostream& operator<<(std::ostream& os, const RayTracer& rayTracer) {
     os << "RayTracer("
